Module Api.Projects

This module manages API parts related to projects.

val get_projects : t -> Equinoxe.Json.t

get_projects t returns all projects associated with the token.

val get_projects_id : t -> id:string -> unit -> Equinoxe.Json.t

get_projects_id t ~id () returns the Json.t that is referenced by the id given in parameter.

val get_projects_id_devices : t -> id:string -> unit -> Equinoxe.Json.t

get_projects_id_devices t ~id () returns the Json.t that contains all the devices related to the project id.

val post_projects_id_devices : t -> id:string -> config:Devices.config -> unit -> Equinoxe.Json.t

post_projects_id_devicest ~id ~config () creates a machine on the Equinix with the Devices.config specification