Module Equinoxe_cohttp.Orga

A module to interact with Equinix organization.

type id

The unique identifier for the an organization.

type config = {

id : id;

name : string;

account_id : string;

website : string;

maintenance_email : string;

max_projects : int;

}

Type that represents an organization configuration.

val id_of_string : string -> id

id_of_string str creates an id from a string from the Equinix API.

val to_string : config -> string

to_string config returns a string representing an organization configuration.

val get_from : t -> id -> config io

get_from t id returns an organization configuration associated with the id given.

val get_all : t -> config list io

get_all t return all the organization associated with the t api token.

val pp : config -> unit

pp config pretty-prints an organization configuration.